+++ src/drivers/driver_w5500.h
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: (c) 2024 Shawn Silverman <shawn@pobox.com>
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+// driver_w5500.h defines things for W5500 boards.
+// This file is part of the QNEthernet library.
+
+#pragma once
+
+#define MTU           1500
+#define MAX_FRAME_LEN 1522  /* Includes the 4-byte FCS (frame check sequence) */